fixed a couple things including tilt slamming, should be 100% ready for posistions and lm2!+1 driver code now. does anyone even read these...
